Move SubscribeToMoreOptions to namespace
